backend/requirements.txt
fastapi
uvicorn
supabase
google-generativeai
python-dotenv
twilio
pydantic
websockets
